fix(node): Remove circular dependency
When running madge on `@sentry/node`:

$ madge --circular src/index.ts
Processed 10 files (698ms) (8 warnings)

✖ Found 1 circular dependency!

1) backend.ts > parsers.ts

To remove the circular relationship between backend and parsers,
we extract the NodeOptions type into a separate types.ts file.
